requirements.txt
aiogram~=3.13.1
python-dotenv~=1.0.1
openai-whisper~=20240927
pydub~=0.25.1
python-docx~=1.1.2